1st program almost done (w/ tf_idf and comments)

+++ source/document_info.c
@@ -3,7 +3,7 @@
 struct DocumentInfo {
     int word_index;
     int word_frequency;
-    int tf_idf;
+    float tf_idf;
 };
 
 DocumentInfo* AllocateDocumentsInfo() {
